requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.43.0
pydantic==2.9.2
pydantic-settings==2.5.2
atlassian-python-api==3.41.16
GitPython==3.1.43
httpx==0.27.2
streamlit==1.39.0
PyGithub==2.3.0
python-dotenv==1.0.1
sse-starlette==2.1.3
